backend/requirements.txt
flask
flask-cors
transformers
faiss-cpu
sentence-transformers
